worker: add untar.go
